Release 3.9.0

+++ src/app/download-manager/downloads-tab/downloads-tab.component.ts
@@ -1,61 +1,59 @@
-import { Component, Input, EventEmitter, Output, NgZone } from '@angular/core';
-import { ContentType, MimeType, MenuOverflow, RouterLinks } from '@app/app/app.constant';
-// import { MenuOverflow } from '../../../app/app.constant';
-import { OverflowMenuComponent } from '@app/app/profile/overflow-menu/overflow-menu.component';
-// import { ViewController } from 'ionic-angular/navigation/view-controller';
-import { CommonUtilService, } from '@app/services/common-util.service';
-import { TelemetryGeneratorService } from '@app/services/telemetry-generator.service';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { MenuOverflow } from '@app/app/app.constant';
 import { SbPopoverComponent } from '@app/app/components/popups/sb-popover/sb-popover.component';
-import { PopoverController, Events } from '@ionic/angular';
-import { InteractType, TelemetryObject } from 'sunbird-sdk';
-import { Content, ContentDelete } from 'sunbird-sdk';
+import { OverflowMenuComponent } from '@app/app/profile/overflow-menu/overflow-menu.component';
+import { AppHeaderService } from '@app/services';
+import { CommonUtilService } from '@app/services/common-util.service';
+import { NavigationService } from '@app/services/navigation-handler.service';
+import { TelemetryGeneratorService } from '@app/services/telemetry-generator.service';
+import { ContentUtil } from '@app/util/content-util';
+import { PopoverController } from '@ionic/angular';
+import { Events } from '@app/util/events';
+import { Content, ContentDelete, CorrelationData, InteractType, TelemetryObject } from 'sunbird-sdk';
+import { ActionButtonType, CorReleationDataType, Environment, InteractSubtype, PageId } from '../../../services/telemetry-constants';
 import { SbGenericPopoverComponent } from '../../components/popups/sb-generic-popover/sb-generic-popover.component';
-import { InteractSubtype, Environment, PageId, ActionButtonType } from '../../../services/telemetry-constants';
 import { EmitedContents } from '../download-manager.interface';
-import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-downloads-tab',
   templateUrl: './downloads-tab.component.html',
   styleUrls: ['./downloads-tab.component.scss'],
 })
-export class DownloadsTabComponent {
+export class DownloadsTabComponent implements OnInit {
 
   @Input() downloadedContents: Content[] = [];
   @Output() deleteContents = new EventEmitter();
   @Output() sortCriteriaChanged = new EventEmitter();
   showLoader = false;
   selectedContents: ContentDelete[] = [];
-  showDeleteButton: Boolean = true;
-  deleteAllPopupPresent: Boolean = false;
-  showSelectAll: boolean = true;
+  showDeleteButton = true;
+  deleteAllPopupPresent = false;
+  showSelectAll = true;
   selectedFilter: string = MenuOverflow.DOWNLOAD_FILTERS[0];
   deleteAllConfirm;
   selectedContentsInfo = {
     totalSize: 0,
     count: 0
   };
+  defaultImg = this.commonUtilService.convertFileSrc('assets/imgs/ic_launcher.png');
 
   constructor(
     private popoverCtrl: PopoverController,
     private commonUtilService: CommonUtilService,
     private events: Events,
     private telemetryGeneratorService: TelemetryGeneratorService,
-    private router: Router,
-    private zone: NgZone) {
-    this.setSelectedItems();
+    private navService: NavigationService,
+    private headerService: AppHeaderService) {
   }
 
-  setSelectedItems() {
-    this.downloadedContents.forEach(element => {
-      element['isSelected'] = false;
+  ngOnInit(): void {
+    this.headerService.headerEventEmitted$.subscribe(async () => {
+      if (this.deleteAllPopupPresent) {
+        await this.deleteAllConfirm.dismiss();
+      }
     });
-    console.log('contents', this.downloadedContents);
   }
 
-  ngOnChanges() {
-    this.setSelectedItems();
-  }
   async showDeletePopup(identifier?) {
     if (identifier) {
       this.telemetryGeneratorService.generateInteractTelemetry(InteractType.TOUCH,
@@ -81,43 +79,47 @@ export class DownloadsTabComponent {
           },
         ],
         icon: null,
-        // mshowDeletePopupshowDeletePopupetaInfo: this.content.contentData.name,
         sbPopoverContent: identifier ? this.commonUtilService.translateMessage('DELETE_CONTENT_WARNING')
           : this.commonUtilService.translateMessage('DELETE_ALL_CONTENT_WARNING')
       },
       cssClass: 'sb-popover danger',
     });
     await deleteConfirm.present();
-    const response = await deleteConfirm.onDidDismiss();
-    console.log('downloads tab', response);
-    switch (response.data) {
-      case undefined:
-        this.unSelectAllContents();
-        this.telemetryGeneratorService.generateInteractTelemetry(
-          InteractType.TOUCH,
-          InteractSubtype.CLOSE_CLICKED,
-          Environment.DOWNLOADS,
-          PageId.SINGLE_DELETE_CONFIRMATION_POPUP);
-        break;
-      case null:
-        this.unSelectAllContents();
-        this.telemetryGeneratorService.generateInteractTelemetry(
-          InteractType.TOUCH,
-          InteractSubtype.OUTSIDE_POPUP_AREA_CLICKED,
-          Environment.DOWNLOADS,
-          PageId.SINGLE_DELETE_CONFIRMATION_POPUP);
-        break;
-      default:
-        const valuesMap = {};
-        valuesMap['type'] = ActionButtonType.POSITIVE;
-        this.telemetryGeneratorService.generateInteractTelemetry(
-          InteractType.TOUCH,
-          InteractSubtype.ACTION_BUTTON_CLICKED,
-          Environment.DOWNLOADS,
-          PageId.SINGLE_DELETE_CONFIRMATION_POPUP, undefined,
-          valuesMap);
-        this.deleteContent();
-        break;
+    const { data } = await deleteConfirm.onDidDismiss();
+
+    if (data === undefined) { // Backdrop clicked
+      if (!identifier) { this.unSelectAllContents(); }
+      this.telemetryGeneratorService.generateInteractTelemetry(
+        InteractType.TOUCH,
+        InteractSubtype.OUTSIDE_POPUP_AREA_CLICKED,
+        Environment.DOWNLOADS,
+        PageId.SINGLE_DELETE_CONFIRMATION_POPUP);
+    } else if (data.closeDeletePopOver) { // Close clicked
+      if (!identifier) { this.unSelectAllContents(); }
+      this.telemetryGeneratorService.generateInteractTelemetry(
+        InteractType.TOUCH,
+        InteractSubtype.CLOSE_CLICKED,
+        Environment.DOWNLOADS,
+        PageId.SINGLE_DELETE_CONFIRMATION_POPUP);
+    } else if (data.canDelete) {
+      const valuesMap = {};
+      valuesMap['type'] = ActionButtonType.POSITIVE;
+      let telemetryObject: TelemetryObject;
+      if (identifier) {
+        this.downloadedContents.forEach(element => {
+          if (element.identifier === identifier) {
+            telemetryObject = ContentUtil.getTelemetryObject(element);
+          }
+        });
+      }
+      this.telemetryGeneratorService.generateInteractTelemetry(
+        InteractType.TOUCH,
+        InteractSubtype.ACTION_BUTTON_CLICKED,
+        Environment.DOWNLOADS,
+        identifier ? PageId.SINGLE_DELETE_CONFIRMATION_POPUP : PageId.BULK_DELETE_CONFIRMATION_POPUP,
+        telemetryObject,
+        valuesMap);
+      this.deleteContent();
     }
   }
 
@@ -151,51 +153,57 @@ export class DownloadsTabComponent {
         this.sortCriteriaChanged.emit(data);
       }
       if (this.deleteAllPopupPresent) {
-        this.deleteAllConfirm.dismiss(null);
+        await this.deleteAllConfirm.dismiss({ isLeftButtonClicked: null });
       }
     }
   }
 
   selectAllContents() {
-    this.zone.run(() => {
-      this.downloadedContents.forEach(element => {
-        element['isSelected'] = true;
-      });
-      this.showDeleteButton = false;
-      this.showSelectAll = false;
-      this.deleteAllContents();
+    this.telemetryGeneratorService.generateInteractTelemetry(
+      InteractType.TOUCH,
+      InteractSubtype.SELECT_ALL_CLICKED,
+      Environment.DOWNLOADS,
+      PageId.DOWNLOADS);
+    this.downloadedContents.forEach(element => {
+      element['isSelected'] = true;
     });
+    this.showDeleteButton = false;
+    this.showSelectAll = false;
+    this.deleteAllContents();
   }
 
-  unSelectAllContents(event?) {
+  async unSelectAllContents(event?) {
+    this.telemetryGeneratorService.generateInteractTelemetry(
+      InteractType.TOUCH,
+      InteractSubtype.UNSELECT_ALL_CLICKED,
+      Environment.DOWNLOADS,
+      PageId.DOWNLOADS);
     this.downloadedContents.forEach(element => {
       element['isSelected'] = false;
     });
-
     this.showDeleteButton = true;
     this.showSelectAll = true;
     if (this.deleteAllPopupPresent) {
-      this.deleteAllConfirm.dismiss(null);
+      await this.deleteAllConfirm.dismiss({ isLeftButtonClicked: null });
     }
     this.selectedContents = [];
   }
 
   async toggleContentSelect(event, idx) {
-    // this.downloadedContents[idx]['isSelected'] = !this.downloadedContents[idx]['isSelected'];
-    if (event.value) {
-      this.downloadedContents[idx]['isSelected'] = event.value;
-      const selectedContents = (this.downloadedContents.filter((element) => element['isSelected']));
-      if (selectedContents.length) {
-        if (selectedContents.length === this.downloadedContents.length) {
-          this.showSelectAll = false;
-        } else {
-          this.showSelectAll = true;
-        }
-        this.showDeleteButton = false;
-        this.deleteAllContents();
+    this.downloadedContents[idx]['isSelected'] = event.detail.checked;
+    const selectedContents = (this.downloadedContents.filter((element) => element['isSelected']));
+    if (selectedContents.length) {
+      if (selectedContents.length === this.downloadedContents.length) {
+        this.showSelectAll = false;
       } else {
-        this.showDeleteButton = true;
-        await this.deleteAllConfirm.dismiss(null);
+        this.showSelectAll = true;
+      }
+      this.showDeleteButton = false;
+      this.deleteAllContents();
+    } else {
+      this.showDeleteButton = true;
+      if (this.deleteAllPopupPresent) {
+        await this.deleteAllConfirm.dismiss({ isLeftButtonClicked: null });
       }
     }
   }
@@ -222,6 +230,7 @@ export class DownloadsTabComponent {
     });
     if (!this.deleteAllPopupPresent) {
       this.telemetryGeneratorService.generatePageViewTelemetry(PageId.BULK_DELETE_POPUP, Environment.DOWNLOADS);
+      this.deleteAllPopupPresent = true;
       this.deleteAllConfirm = await this.popoverCtrl.create({
         component: SbGenericPopoverComponent,
         componentProps: {
@@ -241,25 +250,23 @@ export class DownloadsTabComponent {
         },
         cssClass: 'sb-popover danger sb-dw-delete-popover',
         showBackdrop: false,
-        // migration-TODO
-        // enableBackdropDismiss: false (ionic v3)
-        backdropDismiss: false
+        backdropDismiss: false,
+        animated: true
       });
       await this.deleteAllConfirm.present();
-      this.deleteAllPopupPresent = true;
-    }
-    await this.deleteAllConfirm.onDidDismiss((leftBtnClicked: any) => {
+      const { data } = await this.deleteAllConfirm.onDidDismiss();
       this.deleteAllPopupPresent = false;
       const valuesMap = {};
-      if (leftBtnClicked == null) {
+      if (data && data.isLeftButtonClicked === null) {
         this.unSelectAllContents();
         this.telemetryGeneratorService.generateInteractTelemetry(
           InteractType.TOUCH,
           InteractSubtype.POPUP_DISMISSED,
           Environment.DOWNLOADS,
           PageId.BULK_DELETE_POPUP);
+        this.deleteAllConfirm = undefined;
         return;
-      } else if (leftBtnClicked) {
+      } else if (data.isLeftButtonClicked) {
         valuesMap['type'] = ActionButtonType.NEGATIVE;
         this.unSelectAllContents();
       } else {
@@ -272,48 +279,31 @@ export class DownloadsTabComponent {
           valuesMap);
         this.showDeletePopup();
       }
+      this.deleteAllConfirm = undefined;
       this.telemetryGeneratorService.generateInteractTelemetry(
         InteractType.TOUCH,
         InteractSubtype.ACTION_BUTTON_CLICKED,
         Environment.DOWNLOADS,
         PageId.BULK_DELETE_POPUP, undefined,
         valuesMap);
-    });
+    }
   }
 
   navigateToDetailsPage(content) {
-    const objectType = this.telemetryGeneratorService.isCollection(content.mimeType) ? content.contentData.contentType
-      : ContentType.RESOURCE;
-    const telemetryObject: TelemetryObject = new TelemetryObject(content.identifier, objectType, content.contentData.pkgVersion);
+    const corRelationList: Array<CorrelationData> = [{
+        id: CorReleationDataType.DOWNLOADS,
+        type: CorReleationDataType.SECTION
+      }];
     this.telemetryGeneratorService.generateInteractTelemetry(InteractType.TOUCH,
       InteractSubtype.CONTENT_CLICKED,
       Environment.DOWNLOADS,
       PageId.DOWNLOADS,
-      telemetryObject);
-    if (!this.selectedContents.length) {
-      switch (content.mimeType) {
-
-        // case MimeType.COLLECTION: this.navCtrl.push(CollectionDetailsEtbPage, { content: content });
-        case MimeType.COLLECTION: this.router.navigate([RouterLinks.COLLECTION_DETAIL_ETB], {
-          state: {
-            content: content
-          }
-        });
-
-          break;
-        // default: this.navCtrl.push(ContentDetailsPage, { content: content });
-        default: this.router.navigate([RouterLinks.CONTENT_DETAILS], {
-          state: {
-            content: content
-          }
-        });
-      }
-    }
-
+      ContentUtil.getTelemetryObject(content),
+      undefined,
+      ContentUtil.generateRollUp(undefined, content.identifier),
+      corRelationList);
+    this.navService.navigateToDetailPage(
+      content, { content }
+    );
   }
-
 }
-
-
-
-
