+++ src/services/content/content-delete-handler.ts
@@ -1,5 +1,6 @@
 import { Injectable, Inject } from '@angular/core';
-import { PopoverController, Events } from '@ionic/angular';
+import { PopoverController } from '@ionic/angular';
+import { Events } from '@app/util/events';
 import { Subject } from 'rxjs';
 import { ContentService, InteractType, Content, ContentDeleteStatus } from 'sunbird-sdk';
 
@@ -44,7 +45,7 @@ export class ContentDeleteHandler {
                 content,
                 isChild: isChildContent,
                 objRollup: contentInfo.rollUp,
-                pageName: PageId.CONTENT_DETAIL,
+                pageName: pageId,
                 corRelationList: contentInfo.correlationList,
                 sbPopoverHeading: this.commonUtilService.translateMessage('DELETE'),
                 sbPopoverMainTitle: this.commonUtilService.translateMessage('CONTENT_DELETE'),
@@ -56,13 +57,14 @@ export class ContentDeleteHandler {
                 ],
                 icon: null,
                 metaInfo: content.contentData.name,
-                sbPopoverContent: ' 1 item' + ' (' + this.fileSizePipe.transform(content.sizeOnDevice, 2) + ')',
+                sbPopoverContent: this.commonUtilService.translateMessage('DELETE_CONTENT_SIZE',
+                {content_size: this.fileSizePipe.transform(content.sizeOnDevice ? content.sizeOnDevice : content.contentData.size, 2)})
             },
             cssClass: 'sb-popover danger',
         });
         await confirm.present();
         const { data } = await confirm.onDidDismiss();
-        if (data) {
+        if (data && data.canDelete) {
             this.deleteContent(content.identifier, isChildContent, contentInfo, pageId);
         }
     }
@@ -70,7 +72,7 @@ export class ContentDeleteHandler {
     /**
      * Deletes the content
      */
-    private async deleteContent(identifier: string, isChildContent: boolean, contentInfo: ContentInfo, pageId: string) {
+    public async deleteContent(identifier: string, isChildContent: boolean, contentInfo: ContentInfo, pageId: string) {
         this.telemetryGeneratorService.generateInteractTelemetry(
             InteractType.TOUCH,
             InteractSubtype.DELETE_CLICKED,
@@ -100,7 +102,7 @@ export class ContentDeleteHandler {
                 this.contentDelete.next();
                 this.commonUtilService.showToast('MSG_RESOURCE_DELETED');
             }
-        }).catch(async(error: any) => {
+        }).catch(async (error: any) => {
             await loader.dismiss();
             console.log('delete response: ', error);
             this.commonUtilService.showToast('CONTENT_DELETE_FAILED');
